Make the TrajectorySequenceRunner Super
